+++ enviral/__init__.py
@@ -122,7 +122,7 @@ def serialize(
                             logger.warning(
                                 f"Error attempting to load {key} with value {value}"
                             )
-                    elif schema_type == "array" and "," in value:
+                    elif schema_type == "array":
                         value = [v.strip() for v in value.split(",")]
                 if schema_type == "boolean":
                     value = value.lower() in ("1", "y", "yes", "true")