+++ tests/test_serialize.py
@@ -106,3 +106,14 @@ def test_array_from_comma_values(env):
         }
     )
     assert settings == {"foo": ["foo", "bar"]}
+
+
+def test_array_from_comma_single_value(env):
+    os.environ["FOO"] = "foo"
+    settings = enviral.serialize(
+        {
+            "type": "object",
+            "properties": {"foo": {"type": "array", "items": {"type": "string"}}},
+        }
+    )
+    assert settings == {"foo": ["foo"]}