+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 1.0.4 (unreleased)
 ------------------
-
-- Nothing changed yet.
+- Update serialize to support single item in array
 
 
 1.0.3 (2019-09-29)